+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/adapter/HandlerExecutionAdapter.java
@@ -1,7 +1,7 @@
 package com.interface21.webmvc.servlet.mvc.adapter;
 
-import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.mvc.execution.HandlerExecution;
+import com.interface21.webmvc.servlet.view.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
