+++ mvc/src/main/java/com/interface21/webmvc/servlet/view/View.java
@@ -1,8 +1,7 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.webmvc.servlet.view;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-
 import java.util.Map;
 
 public interface View {
